+++ src/EVault/shared/types/MarketStorage.sol
@@ -7,8 +7,8 @@ import {LTVConfig} from "./LTVConfig.sol";
 import {UserStorage} from "./UserStorage.sol";
 
 struct MarketStorage {
-    // Packed slot 5 + 14 + 2 + 2 + 4 + 1 = 28
-    uint40 lastInterestAccumulatorUpdate;
+    // Packed slot 6 + 14 + 2 + 2 + 4 + 1 = 29
+    uint48 lastInterestAccumulatorUpdate;
     Assets cash;
     AmountCap supplyCap;
     AmountCap borrowCap;
